backend/requirements.txt
# Direct dependencies with versions from requirements.lock
# Install: pip install -r requirements.txt
# Reproducible: pip install -r requirements.lock

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0
numpy==2.4.1

# Web
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22

# LangChain & LangGraph
langchain==1.2.7
langchain-core==1.2.16
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.7
langchain-chroma==1.1.0
langchain-google-genai==4.2.1

# LLM & embeddings
mistralai==1.12.4
sentence-transformers==5.2.3

# Database
psycopg2-binary==2.9.11
SQLAlchemy==2.0.46
chromadb==1.5.0
alembic==1.18.4

# Dev / seed
Faker==40.1.2
pypdf==6.7.1
